# Get rid of defer in interator